[DD-439] Regenerate the code snippets spreadsheet (#6548)
* Seperate total and single code

* Update to seperate code

* update names

* filter out JA files

* remove .md and .adoc
